refactor: 도장판 상세 페이지 리팩토링 (구조 분리)
